LayoutTests/imported/w3c/web-platform-tests/webrtc/rtcpeerconnection/rtcpeerconnection-idl-expected.txt

CONSOLE MESSAGE: line 100: TypeError: RTCPeerConnection argument must be a valid dictionary
Description

This test verifies the availability of the RTCPeerConnection interface.


Harness Error (FAIL), message = TypeError: RTCPeerConnection argument must be a valid dictionary
